app/requirements.txt
scipy
matplotlib
statsmodels
flask
Flask-Bootstrap4
Flask-WTF
flask-limiter
jplephem
ephem
pyopenssl
pyqt5